+++ R/mol_parameters.R
@@ -1316,52 +1316,87 @@ get_cit_paras <- function(lw = NULL, lg = 0, ...) {
   paras
 }
 
+get_atp_31p_paras <- function(lw = NULL, lg = 0, ...) {
+  if (is.null(lw)) lw = 2
+  nucleus <- rep("1H", 3)
+  chem_shift <- c(-7.616, -16.26, -2.6)
+  j_coupling_mat <- matrix(0, 3, 3)
+  j_coupling_mat[2,1] <- 16.3
+  j_coupling_mat[3,2] <- 16.1
+  
+  spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
+                       j_coupling_mat = j_coupling_mat, scale_factor = 1,
+                       lw = lw, lg = lg)
+  
+  source <- "Unknown"
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "ATP", source = source,
+                full_name = "Adenosine triphosphate")
+  
+  class(paras) <- "mol_parameters"
+  paras
+}
 
+get_gpe_31p_paras <- function(lw = NULL, lg = 0, ...) {
+  if (is.null(lw)) lw = 2
+  paras <- get_uncoupled_mol("GPE", 3.5, "1H", 1, lw, lg,
+                             "Glycerol phosphorylethanolamine")
+  
+  source <- "Unknown"
+  
+  paras$source <- source
+  paras
+}
 
-# get_10spin_paras <- function(lw = NULL, lg = 0) {
-#   if (is.null(lw)) lw = 2
-#   nucleus <- rep("1H", 10)
-#   chem_shift <- c(5.216, 3.519, 3.698, 3.395, 3.822, 3.826, 3.749, 1, 2, 3)
-#   j_coupling_mat <- matrix(0, 10, 10)
-#   j_coupling_mat[2,1] <- 3.8
-#   j_coupling_mat[3,2] <- 9.6
-#   j_coupling_mat[4,3] <- 9.4
-#   j_coupling_mat[5,4] <- 9.9
-#   j_coupling_mat[6,5] <- 1.5
-#   j_coupling_mat[7,5] <- 6
-#   j_coupling_mat[7,6] <- -12.1
-#   
-#   spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
-#                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
-#                        lw = lw, lg = lg)
-#   
-#   source <- "made up molecule"
-#   
-#   paras <- list(spin_groups = list(spin_group_a), name = "Glc", source = source)
-#   class(paras) <- "mol_parameters"
-#   paras
-# }
+get_nadh_31p_paras <- function(lw = NULL, lg = 0, ...) {
+  if (is.null(lw)) lw = 2
+  paras <- get_uncoupled_mol("NADH", -8.3, "1H", 1, lw, lg,
+                             "Nicotinamide adenine dinucleotide, reduced")
+  
+  source <- "Unknown"
+  
+  paras$source <- source
+  paras
+}
 
-# get_9spin_paras <- function(lw = NULL, lg = 0) {
-#   if (is.null(lw)) lw = 2
-#   nucleus <- rep("1H", 9)
-#   chem_shift <- c(5.216, 3.519, 3.698, 3.395, 3.822, 3.826, 3.749, 1, 2)
-#   j_coupling_mat <- matrix(0, 9, 9)
-#   j_coupling_mat[2,1] <- 3.8
-#   j_coupling_mat[3,2] <- 9.6
-#   j_coupling_mat[4,3] <- 9.4
-#   j_coupling_mat[5,4] <- 9.9
-#   j_coupling_mat[6,5] <- 1.5
-#   j_coupling_mat[7,5] <- 6
-#   j_coupling_mat[7,6] <- -12.1
-#   
-#   spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
-#                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
-#                        lw = lw, lg = lg)
-#   
-#   source <- "made up molecule"
-#   
-#   paras <- list(spin_groups = list(spin_group_a), name = "Glc", source = source)
-#   class(paras) <- "mol_parameters"
-#   paras
-# }
+get_pch_31p_paras <- function(lw = NULL, lg = 0, ...) {
+  if (is.null(lw)) lw = 2
+  paras <- get_uncoupled_mol("PCh", 6.243, "1H", 1, lw, lg, "Phosphorylcholine")
+  
+  source <- "Unknown"
+  
+  paras$source <- source
+  paras
+}
+
+get_pcr_31p_paras <- function(lw = NULL, lg = 0, ...) {
+  if (is.null(lw)) lw = 2
+  paras <- get_uncoupled_mol("PCr", 0, "1H", 1, lw, lg, "Phosphocreatine")
+  
+  source <- "Unknown"
+  
+  paras$source <- source
+  paras
+}
+
+get_pe_31p_paras <- function(lw = NULL, lg = 0, ...) {
+  if (is.null(lw)) lw = 2
+  paras <- get_uncoupled_mol("PE", 6.7222, "1H", 1, lw, lg,
+                             "Phosphorylethanolamine ")
+  
+  source <- "Unknown"
+  
+  paras$source <- source
+  paras
+}
+
+get_pi_31p_paras <- function(lw = NULL, lg = 0, ...) {
+  if (is.null(lw)) lw = 2
+  paras <- get_uncoupled_mol("PI", 4.8161, "1H", 1, lw, lg,
+                             "Inorganic phosphate")
+  
+  source <- "Unknown"
+  
+  paras$source <- source
+  paras
+}