+++ NEWS.md
@@ -1,7 +1,7 @@
 # spant 2.16.0
 * Added subset option to mean_dyns function.
 * Added vline option to plot.
-* Added typical 31P brain metabolites.
+* Added common 31P brain metabolites.
 
 # spant 2.15.0
 * An SNR estimate of zero in comb_coils is now a failure case.